+++ Pandora/Core/Morphism.hs
@@ -1,7 +1,6 @@
-module Pandora.Core.Morphism (fix, (&), (%)) where
+module Pandora.Core.Morphism (fix, (&)) where
 
 infixl 1 &
-infixr 9 %
 
 fix :: (a -> a) -> a
 fix f = let x = f x in x
@@ -9,7 +8,3 @@ fix f = let x = f x in x
 {-# INLINE (&) #-}
 (&) :: a -> (a -> b) -> b
 x & f = f x
-
-{-# INLINE (%) #-}
-(%) :: (a -> b -> c) -> b -> a -> c
-(%) f x y = f y x