+++ Pandora/Paradigm/Primary/Functor/Function.hs
@@ -2,7 +2,6 @@
 
 module Pandora.Paradigm.Primary.Functor.Function where
 
-import Pandora.Core.Morphism ((%))
 import Pandora.Pattern.Category ((.), identity)
 import Pandora.Pattern.Functor.Covariant (Covariant ((<$>)))
 import Pandora.Pattern.Functor.Applicative (Applicative ((<*>)))
@@ -12,6 +11,7 @@ import Pandora.Pattern.Functor.Bindable (Bindable ((>>=)))
 import Pandora.Pattern.Functor.Representable (Representable (Representation, (<#>), tabulate))
 
 infixr 2 !
+infixr 9 %
 
 instance Covariant ((->) a) where
 	(<$>) = (.)
@@ -36,3 +36,7 @@ instance Representable ((->) e) where
 {-# INLINE (!) #-}
 (!) :: a -> b -> a
 x ! _ = x
+
+{-# INLINE (%) #-}
+(%) :: (a -> b -> c) -> b -> a -> c
+(%) f x y = f y x